Add support for multiple gases when using Brie
Previous behavior was a single gas constituent as the "gas" component
in Brie mixing. This was by default the first encountered "gas"
constituent or the first constituent if no "gas" was present, but
could be overridden with the "index_of_gas" keyword. Now, a list of
indices can instead be specified by the "brie_gases" keyword. If the
keyword is not provided, "carbon_dioxide" and "gas" type constituents
are mixed according to Wood, before being used as the gas component
of Brie.