Fix error when mixing partially vectorized fluids
When using Wood's mixing model, saturations may be provided as arrays
or scalars. However, a combination of the two will cause an error
when constructing a Material. The error is related to how np.array
handles a list of a combination of arrays and scalars, and a
related deprecation warning is produced by
Fluids.fractions_should_not_exceed_one. Using built-in 'sum' should
solve this issue, instead of np.array followed by np.sum.