Set coordination number for cemented models to be constant and independent of coordination number for friable models
Even if this parameter has the same name for constant cement model, contact cement model and friable model, it is only the friable model that uses a coordination number that is a function of porosity. Cemented models require a fixed value for this parameter, and the original setting is 9.0, and that is also what has been used in other implementations of the models.